+++ CHANGELOG.md
@@ -2,15 +2,25 @@
 
 All notable changes to the dcli tool project will be documented in this file.
 
-See [DSH Api Client Changelog](dsh_api/CHANGELOG.md) for the changelog for the DSH Api Client.
-
-
 ## [Unreleased]
 
 ### Security
 
 * At this time the api credentials must be configured by specifying environment variables. 
   This must be replaced by a more secure (encrypted) solution.
+
+## [0.2.0]
+
+### Added
+
+* New subject for exposed metrics.
+* Save settings in application directory.
+* Capability to generate autocomplete files.
+
+### Removed
+
+The dsh-api subproject is removed and is now in its own repo at 
+[github.com/kpn-dsh/dsh-api](https://github.com/kpn-dsh/dsh-api).
 
 ## [0.1.0] - 2024-10-29